Restructured controllers to internal folder
